gaus/requirements/production.txt
Django
South
factory-boy
Unipath
gunicorn
django-braces
django-spurl
mysql-connector-python==1.2.2
django-extensions==1.3.8
PyMySQL==0.6.2
djangorestframework==2.3.14

